Allow activities outside of swfr. Moved the activities to the examples folder. Moved shell and output to lib, now module exports of index.js so that users can use them to build custom activities.
