+++ examples/activities/echo.js
