+++ examples/activities/list_bucket.js
