+++ lib/activities.js
@@ -3,9 +3,14 @@
 var fs = require("fs"),
     path = require("path");
 
-var activities = fs.readdirSync(path.join(__dirname, "activities"))
+var activities = function activities(activitiesFolder) {
+  return fs.readdirSync(activitiesFolder)
+  .filter(function(a) { 
+    var suffix = "js";
+    return a.indexOf(suffix, a.length - suffix.length) !== -1;
+  })
   .map(function(a) {
-    var activity = require(path.join(__dirname, "activities", a));
+    var activity = require(path.join(activitiesFolder, a));
 
     if (activity.name && activity.version) {
       return [JSON.stringify({
@@ -24,10 +29,13 @@ var activities = fs.readdirSync(path.join(__dirname, "activities"))
 
     return hash;
   }, {});
+};
 
-module.exports = function(name, version) {
-  return activities[JSON.stringify({
-    name: name,
-    version: version
-  })];
+module.exports = function(activitiesFolder) {
+  return function(name, version) {
+    return activities(activitiesFolder)[JSON.stringify({
+      name: name,
+      version: version
+    })];
+  };
 };
