+++ lib/output.js
