+++ lib/s3_output.js
