+++ lib/shell.js
