+++ lib/sync-decider-context.js
@@ -1,15 +1,19 @@
 "use strict";
 
+var path = require("path"), 
+    util = require("util");
+
 var debug = require("debug")("swfr:decider_context"),
     Promise = require("bluebird"),
     humanize = require("humanize-plus");
 
-var lookup = require("./activities");
+var activities = require("./activities");
 
-var SyncDeciderContext = function(task) {
+var SyncDeciderContext = function(task, activitiesFolder) {
   this.task = task;
   this.result = null;
   this.userData = {};
+  this.lookup = activities(activitiesFolder);
 };
 
 SyncDeciderContext.prototype.activity = function() {
@@ -23,10 +27,10 @@ SyncDeciderContext.prototype.activity = function() {
         context = this;
 
     return new Promise(function(resolve, reject) {
-      var activity = lookup(name, version);
+      var activity = context.lookup(name, version);
 
       if (!activity) {
-        return reject(new Error("Unable to locate activity: %s@%s", name, version));
+        return reject(new Error(util.format("Unable to locate activity: %s@%s", name, version)));
       }
 
       return activity.apply(context, args.concat(function(err, result) {
