+++ lib/sync-decider-worker.js
@@ -7,7 +7,7 @@ var P = require("bluebird");
 
 var SyncDeciderContext = require("./sync-decider-context");
 
-var SyncDeciderWorker = function(fn) {
+var SyncDeciderWorker = function(fn, activitiesFolder) {
   stream.Writable.call(this, {
     objectMode: true,
     highWaterMark: 1 // limit the number of buffered tasks
@@ -15,7 +15,7 @@ var SyncDeciderWorker = function(fn) {
 
   this._write = function(task, encoding, callback) {
     // pass the input to the function and provide the rest as the context
-    var context = new SyncDeciderContext(task),
+    var context = new SyncDeciderContext(task, activitiesFolder),
         chain = P.bind(context),
         input = task.input;
